setting right fan and pressure

+++ src/Pressure.cpp
@@ -2,10 +2,10 @@
 
 Pressure::Pressure() : i2c(i2c_config) { }
 
-int Pressure::getPressure() {
+int16_t Pressure::getPressure() {
     uint8_t data[3];
     uint8_t cmd = READ_COMMAND;
-    int pressure = 0;
+    int16_t pressure = 0;
     i2c.transaction(PRESSURE_ADDRESS, &cmd, 1, data, 3);
     /* i2c.write(PRESSURE_ADDRESS, &cmd, 1); */
     /* i2c.read(PRESSURE_ADDRESS, data, 3); */
@@ -14,6 +14,6 @@ int Pressure::getPressure() {
     return convertPressure(pressure);
 }
 
-int Pressure::convertPressure(int pressure) {
-    return ((double)pressure / SCALE_FACTOR) * 0.95;
+int16_t Pressure::convertPressure(int16_t pressure) {
+    return (pressure / SCALE_FACTOR) * 0.95;
 }
